feat: add support for custom serializers for types
